pull the stringMatch code out of QuickOpen and into a new module
