+++ test/UnitTestSuite.js
@@ -47,7 +47,7 @@ define(function (require, exports, module) {
     require("spec/NativeFileSystem-test");
     require("spec/PreferencesManager-test");
     require("spec/ProjectManager-test");
-    require("spec/QuickOpen-test");
+    require("spec/StringMatch-test");
     require("spec/UpdateNotification-test");
     require("spec/ViewUtils-test");
     require("spec/WorkingSetView-test");
